serata1/pyproject.toml
[tool.poetry]
name = "pbg-demo"
version = "0.0.1"
description = "PythonBiellaGroup Streamlit demo"
authors = ["PBG"]
license = "MIT"
readme = "README.md"
repository = ""
documentation = ""
include = ["LICENSE", "README.md"]
keywords = [
    "Data",
    "AI",
    "Dashboard",
]

[build-system]
requires = ["poetry-core>=1.1.13", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
numpy = "^1.21"
pymongo = "^4.0.1"
PyMySQL = "^1.0.2"
pydantic = "^1.9.0"
pytest = "^7.2.0"
pyaml = "^21.8.3"
pandas = "^1.3.3"
python-dotenv = "^0.21.0"
redis = "^4.1.0"
requests = "^2.26.0"
SQLAlchemy = "^1.3"
sqlalchemy-json = "^0.5.0"
sqlalchemy-stubs = "^0.4"
tqdm = "^4.62.3"
Unidecode = "^1.3.2"
psycopg2-binary = "^2.9.1"
mysql-connector-python = "^8.0.27"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^5"
flake8-isort = "^5.0.0"
flake8-builtins = "^2.0"
flake8-bandit = "^4"
autopep8 = "^2.0"
pylint = "^2.8.2"
bandit = "^1.7.0"
black = "^22"
autoflake = "^1.3.1"
pytest-cov = "^4.0.0"
ipython = "^8.0"
commitizen = "^2.20.0"
jupyter = "^1.0.0"
mypy = "0.990"

# --- ISORT config --------------------------------------------- #
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

# --- PyTest config -------------------------------------=------ #
[tool.pytest.ini_options]
markers = [
    "base: test base functionalities",
]

# --- Commitizen Release ----------------------------------------- #
[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$major.$minor.$patch$prerelease"
version = "0.1.0"
version_files = [
    "data-connector/__version__.py",
    "pyproject.toml:version"
]
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

[mypy]
ignore_missing_imports = true

[[tool.poetry.source]]
    name = "pypi.org"
    url = "https://pypi.org/simple"
    default = true
